Move all display logic to its own module
