+++ compiler/noirc_frontend/src/hir/comptime/mod.rs
@@ -1,7 +1,7 @@
+mod display;
 mod errors;
 mod hir_to_display_ast;
 mod interpreter;
-mod quoted_pretty_printing;
 mod tests;
 mod value;
